pyproject.toml
[project]
name = "missionary-board"
version = "0.1.0"
description = "Missionary Portrait Board"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogoogle>=5.13.2",
    "argon2-cffi>=23.1.0",
    "authlib>=1.4.1",
    "cryptography>=44.0.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.5",
    "python-multipart>=0.0.20",
    "sqlitedict>=2.1.0",
    "starlette>=0.45.3",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "mimesis>=18.0.0",
    "pre-commit>=4.1.0",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.4",
]

[project.scripts]
#mboard = "uvicorn --reload --app-dir src --log-config conf/uvicorn.logger.json mboard.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mboard"]

[tool.pytest.ini_options]
# Allow tests to import application files by updating pythonpath. See:
# https://docs.pytest.org/en/latest/explanation/goodpractices.html#choosing-a-test-layout-import-rules
pythonpath = "src"

# Default fixture scope for aysncio tests.
asyncio_default_fixture_loop_scope = "function"

# Show slowest tests.
addopts = "--durations=5"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN002",   # Missing type annotation for `*args`
    "ANN003",   # Missing type annotation for `**kwargs`
    "COM812",   # Missing trailing comma in a single-line tuple
    "D103",     # Missing docstring in public function
    "D104",     # Missing docstring in public package
    "D107",     # Missing docstring in __init__
    "D203",     # 1 blank line required before class docstring
    "D213",     # Multi-line docstring summary should start at the second line
    "PGH003",   # Use specific rule codes when ignoring type issues
    "RET504",   # Unnecessary variable assignment before `return` statement
    "S105",     # Possible hardcoded password
    "S106",     # Possible hardcoded password
    "TRY300",   # Consider moving this statement to an `else` block
]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["S101"]  # Use of assert detected
